py/confluence-rag/requirements.txt
fastapi
uvicorn
python-dotenv
requests
faiss-cpu
langchain
sentence-transformers